Refactor: start unifying the 2 SlashCommand implementations
