+++ front/components/editor/extensions/shared/slash_suggestion/slashSuggestionUtils.ts
@@ -1,0 +1,40 @@
+import type { Range } from "@tiptap/core";
+import type { EditorState } from "@tiptap/pm/state";
+
+export function hasSlashCharacterAtPosition(
+  state: EditorState,
+  position: number
+) {
+  const docSize = state.doc.content.size;
+
+  if (position < 1 || position > docSize) {
+    return false;
+  }
+
+  return (
+    state.doc.textBetween(
+      position,
+      Math.min(position + 1, docSize + 1),
+      undefined,
+      "\ufffc"
+    ) === "/"
+  );
+}
+
+export function isAllowedSlashQuery(state: EditorState, range: Range) {
+  const text = state.doc.textBetween(range.from, range.to, undefined, "\ufffc");
+
+  if (!text.startsWith("/")) {
+    return false;
+  }
+
+  return !text.slice(1).startsWith(" ");
+}
+
+export function shouldInsertSlashBoundarySpace(state: EditorState) {
+  const textBefore = state.selection.$from.nodeBefore?.isText
+    ? state.selection.$from.nodeBefore.text
+    : null;
+
+  return !!textBefore && !textBefore.endsWith(" ");
+}
